backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask-cors
flask-jwt-extended
python-dotenv
pyjwt
